Create Number Guessing Game - Task 1
Number Guessing Game :

This repository contains my first task for the C++ programming internship: a simple yet engaging Number Guessing Game. The objective of this game is to guess a randomly generated number between 1 and 100.

How It Works :

Random Number Generation: The game uses the C++ standard library's rand() function to generate a random number between 1 and 100. The srand() function is used to seed the random number generator with the current time to ensure different outcomes each time the game is played.

User Interaction: The player is prompted to guess the number. Based on the player's input, the game provides feedback:

If the guess is lower than the random number, the game outputs "Too low. Try again."
If the guess is higher than the random number, the game outputs "Too high. Try again."
Winning Condition: The loop continues until the player correctly guesses the random number, at which point the game congratulates the player with "Congratulations! You guessed the correct number.